requirements.txt
Flask==2.3.2
opencv-python==4.8.0.74
cvzone==1.5.5
python-pptx==0.6.21
pywin32==305
numpy==1.24.2
